fix: rename store.ts
