+++ src/App.tsx
@@ -5,12 +5,12 @@ import LoginPage from './pages/LoginPage.tsx';
 import MainPage from './pages/MainPage.tsx';
 import NotFoundPage from './pages/NotFoundPage.tsx';
 import RegisterPage from './pages/RegisterPage.tsx';
-import { index } from './store';
+import { store } from './store/store.ts';
 import { Header } from './components/Header/Header.tsx';
 
 function App() {
   return (
-    <Provider store={index}>
+    <Provider store={store}>
       <BrowserRouter>
         <Header />
         <Routes>
