+++ src/store/store.ts
@@ -1,10 +1,10 @@
 import { configureStore } from '@reduxjs/toolkit';
 import { userReducer } from '../features/user/userSlice';
 
-export const index = configureStore({
+export const store = configureStore({
   reducer: {
     user: userReducer,
   },
 });
-export type RootState = ReturnType<typeof index.getState>;
-export type AppDispatch = typeof index.dispatch;
+export type RootState = ReturnType<typeof store.getState>;
+export type AppDispatch = typeof store.dispatch;
